[serve] Extract controller actor class factory; make default_impl a leaf seam
Mirror the proxy_actor_class change for the controller: move
get_controller_impl into a new controller_actor_class module that imports
ServeController eagerly. It is only imported by api.py / controller_avatar.py
(never by controller.py), so the import is a clean rung with no cycle.

With both actor-class factories extracted, default_impl no longer imports any
heavy 'up' module at load time and is restored to its original role: a leaf
seam of default factories that the runtime objects (proxy, handle) import. The
load graph is now fully acyclic.

The replica factories (create_replica_impl, create_replica_metrics_manager)
stay as lazy imports: replica.py both defines Replica and calls these factories
on itself, so they cannot be hoisted into an eager rung without splitting
replica.py. Documented inline.
